+++ frontend/components/LiveDocs.js
@@ -1,39 +1,19 @@
-import { html, useState, useRef, useLayoutEffect, useEffect, useMemo, useContext } from "../imports/Preact.js"
-import immer from "../imports/immer.js"
-import observablehq from "../common/SetupCellEnvironment.js"
+import { html, useState, useRef, useEffect } from "../imports/Preact.js"
 import { cl } from "../common/ClassTable.js"
 
-import { RawHTMLContainer, highlight } from "./CellOutput.js"
-import { PlutoActionsContext } from "../common/PlutoContext.js"
-
-const without_workspace_stuff = (str) =>
-    str
-        .replace(/Main\.var&quot;workspace\#\d+&quot;\./g, "") // remove workspace modules from variable names
-        .replace(/Main\.workspace\#\d+\./g, "") // remove workspace modules from variable names
-        .replace(/ in Main\.var&quot;workspace\#\d+&quot;/g, "") // remove workspace modules from method lists
-        .replace(/ in Main\.workspace\#\d+/g, "") // remove workspace modules from method lists
-        .replace(/#&#61;&#61;#[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\:\d+/g, "") // remove UUIDs from filenames
+import { LiveDocsTab } from "./LiveDocsTab.js"
 
 export let LiveDocs = ({ desired_doc_query, on_update_doc_query, notebook }) => {
-    let pluto_actions = useContext(PlutoActionsContext)
     let container_ref = useRef()
-    let live_doc_search_ref = useRef()
-    let [state, set_state] = useState({
-        shown_query: null,
-        searched_query: null,
-        body: "<p>Welcome to the <b>Live docs</b>! Keep this little window open while you work on the notebook, and you will get documentation of everything you type!</p><p>You can also type a query above.</p><hr><p><em>Still stuck? Here are <a href='https://julialang.org/about/help/' target='_blank'>some tips</a>.</em></p>",
-        hidden: true,
-        loading: false,
-    })
-    let update_state = (mutation) => set_state(immer((state) => mutation(state)))
+
+    const [open_tab, set_open_tab] = useState(/** @type { "docs" | "process" | null} */ (null))
+    const hidden = open_tab == null
 
     // Open docs when "open_live_docs" event is triggered
     useEffect(() => {
         let handler = () => {
             // https://github.com/fonsp/Pluto.jl/issues/321
-            update_state((state) => {
-                state.hidden = false
-            })
+            set_open_tab("docs")
             if (window.getComputedStyle(container_ref.current).display === "none") {
                 alert("This browser window is too small to show docs.\n\nMake the window bigger, or try zooming out.")
             }
@@ -42,119 +22,52 @@ export let LiveDocs = ({ desired_doc_query, on_update_doc_query, notebook }) => 
         return () => window.removeEventListener("open_live_docs", handler)
     }, [])
 
-    // Apply syntax highlighting to code blocks:
-    // In the standard HTML container we already do this for code.language-julia blocks,
-    // but in the docs it's safe to extend to to all highlighting I think
-    useLayoutEffect(() => {
-        // Actually, showing the jldoctest stuff wasn't as pretty... should make a mode for that sometimes
-        // for (let code_element of container_ref.current.querySelectorAll("code.language-jldoctest")) {
-        //     highlight(code_element, "julia")
-        // }
-        for (let code_element of container_ref.current.querySelectorAll("code:not([class])")) {
-            highlight(code_element, "julia")
-        }
-    }, [state.body])
-
-    useEffect(() => {
-        if (state.hidden || state.loading) {
-            return
-        }
-        if (!/[^\s]/.test(desired_doc_query)) {
-            // only whitespace
-            return
-        }
-
-        if (state.searched_query !== desired_doc_query) {
-            fetch_docs(desired_doc_query)
-        }
-    }, [desired_doc_query, state.hidden, state.loading, state.searched_query])
-
-    let fetch_docs = (new_query) => {
-        update_state((state) => {
-            state.loading = true
-            state.searched_query = new_query
-        })
-        Promise.race([
-            observablehq.Promises.delay(2000, false),
-            pluto_actions.send("docs", { query: new_query.replace(/^\?/, "") }, { notebook_id: notebook.notebook_id }).then((u) => {
-                if (u.message.status === "⌛") {
-                    return false
-                }
-                if (u.message.status === "👍") {
-                    update_state((state) => {
-                        state.shown_query = new_query
-                        state.body = u.message.doc
-                    })
-                    return true
-                }
-            }),
-        ]).then(() => {
-            update_state((state) => {
-                state.loading = false
-            })
-        })
-    }
-
-    let docs_element = useMemo(() => html` <${RawHTMLContainer} body=${without_workspace_stuff(state.body)} /> `, [state.body])
-    let no_docs_found = state.loading === false && state.searched_query !== "" && state.searched_query !== state.shown_query
-
     return html`
         <aside id="helpbox-wrapper" ref=${container_ref}>
-            <pluto-helpbox class=${cl({ hidden: state.hidden, loading: state.loading, notfound: no_docs_found })}>
-                <header
-                    translate=${false}
-                    onClick=${() => {
-                        if (state.hidden) {
-                            set_state((state) => ({ ...state, hidden: false }))
-                            // wait for next event loop
-                            setTimeout(
-                                () =>
-                                    live_doc_search_ref.current &&
-                                    live_doc_search_ref.current.focus({
-                                        preventScroll: true,
-                                    }),
-                                0
-                            )
-                        }
-                    }}
-                >
-                    ${state.hidden
-                        ? "Live docs"
-                        : html`
-                        <input
-                            title=${no_docs_found ? `"${state.searched_query}" not found` : ""}
-                            id="live-docs-search"
-                            placeholder="Search docs..."
-                            ref=${live_doc_search_ref}
-                            onInput=${(e) => on_update_doc_query(e.target.value)}
-                            value=${desired_doc_query}
-                            type="text"
-                        ></input>
-                        <button onClick=${(e) => {
-                            set_state((state) => ({ ...state, hidden: true }))
-                            e.stopPropagation()
-                            setTimeout(() => live_doc_search_ref.current && live_doc_search_ref.current.focus(), 0)
-                        }}><span></span></button>
-                    `}
+            <pluto-helpbox class=${cl({ hidden, [`helpbox-${open_tab ?? hidden}`]: true })}>
+                <header translate=${false}>
+                    <button
+                        class=${cl({
+                            "helpbox-tab-key": true,
+                            "helpbox-docs": true,
+                            "active": open_tab === "docs",
+                        })}
+                        onClick=${() => {
+                            set_open_tab(open_tab === "docs" ? null : "docs")
+                            // TODO: focus the docs input
+                        }}
+                    >
+                        <span>Live Docs</span>
+                    </button>
+                    <button
+                        class=${cl({
+                            "helpbox-tab-key": true,
+                            "helpbox-process": true,
+                            "active": open_tab === "process",
+                        })}
+                        onClick=${() => {
+                            set_open_tab(open_tab === "process" ? null : "process")
+                        }}
+                    >
+                        <span>Process</span>
+                    </button>
+                    ${hidden
+                        ? null
+                        : html`<button
+                              class="helpbox-close"
+                              onClick=${() => {
+                                  set_open_tab(null)
+                              }}
+                          >
+                              <span></span>
+                          </button>`}
                 </header>
-                <section ref=${(ref) => ref != null && resolve_doc_reference_links(ref, on_update_doc_query)}>
-                    <h1><code>${state.shown_query}</code></h1>
-                    ${docs_element}
-                </section>
+                ${open_tab === "docs"
+                    ? html`<${LiveDocsTab} desired_doc_query=${desired_doc_query} on_update_doc_query=${on_update_doc_query} notebook=${notebook} />`
+                    : open_tab === "process"
+                    ? html`<section>Process TODO</section>`
+                    : null}
             </pluto-helpbox>
         </aside>
     `
-}
-
-const resolve_doc_reference_links = (node, on_update_doc_query) => {
-    for (let anchor of node.querySelectorAll("a")) {
-        const href = anchor.getAttribute("href")
-        if (href != null && href.startsWith("@ref")) {
-            const query = href.length > 4 ? href.substr(5) : anchor.textContent
-            anchor.onclick = (e) => {
-                on_update_doc_query(query)
-                e.preventDefault()
-            }
-        }
-    }
 }
