+++ frontend/dark_color.css
@@ -139,6 +139,7 @@
         --helpbox-bg-color: rgb(30 34 31);
         --helpbox-box-shadow-color: #00000017;
         --helpbox-header-bg-color: #2c3e36;
+        --helpbox-header-tab-bg-color: #554e4e;
         --helpbox-header-color: rgb(255 248 235);
         --helpbox-notfound-header-color: rgb(139, 139, 139);
         --helpbox-text-color: rgb(230, 230, 230);
